Rename integration_tests package to integrationtests
