+++ src/pcap/integrationtests/integration_test.go
@@ -1,4 +1,4 @@
-package integration_tests
+package integrationtests
 
 import (
 	"bytes"
